[#38] Uncomment clojure function. Found problem with LazySeq implemenation.
When using gensym inside a LazySeq, since it will generate a different value each time the LazySeq is realized, different gensym value will be returned if the LazySeq is iterated over more than once. In the loop macro this caused a problem so the LazySeq was converted into a vector.