[#489] Improve Core Erlang code generated for `if` expressions
When functions are inlined (e.g. erlang/<) the ASM generated takes into account
that all the expressions in the `case` are BIFs. Instead of using a low level
test instruction (i.e. `test_ne_eq`) it ends up doing calls to BIFs with the
`gc_bif` or `bif` instruction, which is functionally equivalent but slower.

This implementation copies the same Core Erlang generated by the Erlang compiler.