[#634] Remove 'clj ' prefix to compiler generated core vars
This was originally introduced so that the names wouldn't collide with the
ones generated by the Erlang compiler, but this can never happen since all
Core Erlang code is generated by the clj_emitter.

Also the 'clj ' prefix with the space at the end makes '.core' files impossible
to compile without modifying all the generated var names.

Finally by removing this there is no more concatenation which makes the
compilation times marginally better.